pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "online-media-dl"
version = "0.1.0"
description = "CLI media downloader offline (yt-dlp wrapper) untuk YouTube/IG/TikTok/Facebook/X"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name" }]
license = { text = "MIT" }
dependencies = [
    "yt-dlp>=2024.04.09",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "click>=8.1.3",
    "PyYAML>=6.0.1",
]

[project.scripts]
omdl = "omdl.cli:app"

[tool.setuptools]
packages = ["omdl", "omdl.providers"]
package-dir = { "omdl" = "app/src/omdl" }
include-package-data = true
